Use double for resamplers
Former-commit-id: 6c024e150c3521c9a8f6ac12da96abfd7b4d1be7
Former-commit-id: 598e6e70606a95d16c6205d4eb04a960842246aa
Former-commit-id: 2c3e00cf2db86aaf7eedcbca3b7eef382d689885

+++ src/ImageProcessorCore/Samplers/Resamplers/HermiteResampler.cs
@@ -13,13 +13,13 @@ namespace ImageProcessorCore.Processors
     public class HermiteResampler : IResampler
     {
         /// <inheritdoc/>
-        public float Radius => 2;
+        public double Radius => 2;
 
         /// <inheritdoc/>
-        public float GetValue(float x)
+        public double GetValue(double x)
         {
-            const float B = 0;
-            const float C = 0;
+            const double B = 0;
+            const double C = 0;
 
             return ImageMaths.GetBcValue(x, B, C);
         }
