+++ src/ImageProcessorCore/Samplers/Resamplers/TriangleResampler.cs
@@ -13,10 +13,10 @@ namespace ImageProcessorCore
     public class TriangleResampler : IResampler
     {
         /// <inheritdoc/>
-        public float Radius => 1;
+        public double Radius => 1;
 
         /// <inheritdoc/>
-        public float GetValue(float x)
+        public double GetValue(double x)
         {
             if (x < 0)
             {
